requirements.txt
playwright==1.43.0
python-dotenv==1.0.1
requests==2.31.0
